fix: 에러코드 추가에 따른 서비스코드 일부 수정

+++ src/main/java/com/umc/networkingService/domain/todoList/service/TodoListServiceImpl.java
@@ -39,6 +39,10 @@ public class TodoListServiceImpl implements TodoListService{
     public TodoListIdResponse updateTodoList(Member member, UUID todoListId, TodoListUpdateRequest request){
         ToDoList todoList = todoListRepository.findById(todoListId).orElseThrow(() -> new RestApiException(ErrorCode.EMPTY_TODOLIST));
 
+        if (!todoList.getWriter().getId().equals(member.getId())) {
+            throw new RestApiException(ErrorCode.NO_AUTHORIZATION_TODOLIST);
+        }
+
         todoList.updateTodoList(request.getTitle(), request.getDeadline());
 
         return new TodoListIdResponse((todoList.getId()));
@@ -49,6 +53,10 @@ public class TodoListServiceImpl implements TodoListService{
     public TodoListIdResponse completeTodoList(Member member, UUID todoListId){
         ToDoList todoList = todoListRepository.findById(todoListId).orElseThrow(() -> new RestApiException(ErrorCode.EMPTY_TODOLIST));
 
+        if (!todoList.getWriter().getId().equals(member.getId())) {
+            throw new RestApiException(ErrorCode.NO_AUTHORIZATION_TODOLIST);
+        }
+
         todoList.completeTodoList();
 
         return new TodoListIdResponse(todoList.getId());
@@ -58,6 +66,10 @@ public class TodoListServiceImpl implements TodoListService{
     @Transactional
     public TodoListIdResponse deleteTodoList(Member member, UUID todoListId){
         ToDoList todoList = todoListRepository.findById(todoListId).orElseThrow(() -> new RestApiException(ErrorCode.EMPTY_TODOLIST));
+
+        if (!todoList.getWriter().getId().equals(member.getId())) {
+            throw new RestApiException(ErrorCode.NO_AUTHORIZATION_TODOLIST);
+        }
 
         todoList.delete();
